Fix memory-leak-related issues (#1370)
This PR:

1. fixed a bug where the DP thread will not exit correctly after calling shutdown(). Now it actively throws an interrupt exception instead of waiting for an interruption.
2. fixed memory leak by adding onTerminateDetach on critical observables.
3. Akka creates a lambda function that includes all member variables of AmberClient during the initialization of ClientActor and keeps them forever. So to release all the resources, I separated the initialization of AmberClient and ClientActor. Now ClientActor will first receive an InitializeRequest message which contains necessary workflow information before it receives any other messages.
