+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/worker/promisehandlers/ShutdownDPThreadHandler.scala
@@ -4,8 +4,6 @@ import edu.uci.ics.amber.engine.architecture.worker.WorkerAsyncRPCHandlerInitial
 import edu.uci.ics.amber.engine.architecture.worker.promisehandlers.ShutdownDPThreadHandler.ShutdownDPThread
 import edu.uci.ics.amber.engine.common.rpc.AsyncRPCServer.ControlCommand
 
-import java.util.concurrent.CompletableFuture
-
 object ShutdownDPThreadHandler {
   final case class ShutdownDPThread() extends ControlCommand[Unit]
 }
@@ -16,8 +14,8 @@ trait ShutdownDPThreadHandler {
   registerHandler { (msg: ShutdownDPThread, sender) =>
     {
       dataProcessor.shutdown()
-      new CompletableFuture[Void]().get // wait here to be interrupted
-      () // return unit. this will actually never be called
+      throw new InterruptedException() // actively interrupt itself
+      ()
     }
   }
