+++ core/amber/src/main/scala/edu/uci/ics/texera/web/service/JobRuntimeService.scala
@@ -7,7 +7,6 @@ import edu.uci.ics.amber.engine.architecture.breakpoint.globalbreakpoint.{
   CountGlobalBreakpoint
 }
 import edu.uci.ics.amber.engine.architecture.controller.ControllerEvent._
-import edu.uci.ics.amber.engine.architecture.controller.Workflow
 import edu.uci.ics.amber.engine.architecture.controller.promisehandlers.AssignBreakpointHandler.AssignGlobalBreakpoint
 import edu.uci.ics.amber.engine.architecture.controller.promisehandlers.EvaluatePythonExpressionHandler.EvaluatePythonExpression
 import edu.uci.ics.amber.engine.architecture.controller.promisehandlers.FatalErrorHandler.FatalError
@@ -17,7 +16,8 @@ import edu.uci.ics.amber.engine.architecture.controller.promisehandlers.ResumeHa
 import edu.uci.ics.amber.engine.architecture.controller.promisehandlers.RetryWorkflowHandler.RetryWorkflow
 import edu.uci.ics.amber.engine.architecture.controller.promisehandlers.StartWorkflowHandler.StartWorkflow
 import edu.uci.ics.amber.engine.architecture.principal.{OperatorState, OperatorStatistics}
-import edu.uci.ics.amber.engine.common.{AmberClient, AmberUtils}
+import edu.uci.ics.amber.engine.common.AmberUtils
+import edu.uci.ics.amber.engine.common.client.AmberClient
 import edu.uci.ics.texera.web.SnapshotMulticast
 import edu.uci.ics.texera.web.model.common.FaultedTupleFrontend
 import edu.uci.ics.texera.web.model.websocket.event._
